Reporter log debugger added.
added base class for networkbehaviour
added base class for buildingmanager, tilemanager

changed map transferring method to send dirty tiles
adding packet handler.

+++ Test/Assets/Scripts/Weapons/Rope.cs
@@ -209,11 +209,11 @@ public class Rope : Projectile {
         switch(info.m_RID.m_type)
         {
             case ObjectType.TILE:
-                ropable = TileManager.Inst.Get((int)info.m_RID.m_ID);
+                ropable = TileManager.Inst.Get((ushort)info.m_RID.m_ID);
                 break;
 
             case ObjectType.BUILDING:
-                ropable = BuildingManager.Inst.Get((int)info.m_RID.m_ID);
+                ropable = BuildingManager.Inst.Get((ushort)info.m_RID.m_ID);
                 break;
 
             case ObjectType.PLAYER:
